change logger error interface
