+++ example-app/SantokuApp/src/bases/core/error/RuntimeError.ts
@@ -1,0 +1,7 @@
+import {ErrorWrapper} from './ErrorWrapper';
+
+export class RuntimeError extends ErrorWrapper {}
+
+export function isRuntimeError(error?: any): error is RuntimeError {
+  return error != null && typeof error === 'object' && error instanceof RuntimeError;
+}
