+++ example-app/SantokuApp/src/bases/logging/Transport.ts
@@ -2,7 +2,7 @@ import {LogLevel} from './Logger';
 
 interface Transport {
   log: TransportMethod;
-  error: LeveledTransportMethodWithErrorCode;
+  error: ErroredTransportMethod;
   warn: LeveledTransportMethod;
   info: LeveledTransportMethod;
   debug: LeveledTransportMethod;
@@ -10,15 +10,15 @@ interface Transport {
 }
 
 interface TransportMethod {
-  (level: LogLevel, message: string, errorCode?: string): void;
+  (level: LogLevel, message: string | Error, errorCode?: string): void;
 }
 
 interface LeveledTransportMethod {
   (message: string): void;
 }
 
-interface LeveledTransportMethodWithErrorCode {
-  (message: string, errorCode: string): void;
+interface ErroredTransportMethod {
+  (error: Error, errorCode: string): void;
 }
 
 export type {Transport};
